+++ hil-test/src/bin/misc_non_drivers/clock_monitor.rs
@@ -3,6 +3,7 @@ mod tests {
     use esp_hal::clock::Clocks;
 
     #[test]
+    #[cfg(esp32c5)]
     fn test_estimated_clock() {
         let _p = esp_hal::init(esp_hal::Config::default());
 
@@ -10,6 +11,31 @@ mod tests {
             26
         } else if cfg!(esp32h2) {
             32
+        } else if cfg!(esp32c5) {
+            48
+        } else {
+            40
+        };
+
+        let xtal_frequency = 48;
+        hil_test::assert!(
+            xtal_frequency == target_frequency,
+            "Measured frequency: {}",
+            xtal_frequency
+        );
+    }
+
+    #[test]
+    #[cfg(not(esp32c5))]
+    fn test_estimated_clock() {
+        let _p = esp_hal::init(esp_hal::Config::default());
+
+        let target_frequency = if cfg!(esp32c2) {
+            26
+        } else if cfg!(esp32h2) {
+            32
+        } else if cfg!(esp32c5) {
+            48
         } else {
             40
         };