+++ qa-test/src/bin/embassy_wifi_csi.rs
@@ -1,20 +1,15 @@
-//! This example creates BLE and WiFi and once it gets an IP address the BLE controller is dropped
-//! and shows that WiFi still works afterwards.
+//! CSI Example
 //!
-//! - set SSID and PASSWORD env variable
-//! - gets an ip address via DHCP
-//! - performs an HTTP get request to some "random" server
+//! Set SSID and PASSWORD env variable before running this example.
 
-//% FEATURES: esp-radio esp-radio/wifi esp-radio/ble esp-radio/unstable esp-hal/unstable
-//% CHIPS: esp32 esp32c2 esp32c3 esp32c6 esp32s3
+//% CHIPS: esp32 esp32c2 esp32c3 esp32c6 esp32s2 esp32s3
+//% FEATURES: esp-radio esp-radio/wifi esp-radio/csi esp-radio/unstable esp-hal/unstable
 
 #![no_std]
 #![no_main]
 
-use core::net::Ipv4Addr;
-
 use embassy_executor::Spawner;
-use embassy_net::{Runner, StackResources, tcp::TcpSocket};
+use embassy_net::{Runner, StackResources};
 use embassy_time::{Duration, Timer};
 use esp_alloc as _;
 use esp_backtrace as _;
@@ -26,17 +21,16 @@ use esp_hal::{
     timer::timg::TimerGroup,
 };
 use esp_println::println;
-use esp_radio::{
-    ble::controller::BleConnector,
-    wifi::{
-        ModeConfig,
-        WifiController,
-        WifiDevice,
-        WifiEvent,
-        scan::ScanConfig,
-        sta::StationConfig,
-    },
+use esp_radio::wifi::{
+    ModeConfig,
+    WifiController,
+    WifiDevice,
+    WifiEvent,
+    csi::CsiConfig,
+    scan::ScanConfig,
+    sta::StationConfig,
 };
+
 esp_bootloader_esp_idf::esp_app_desc!();
 
 // When you are okay with using a nightly compiler it's better to use https://docs.rs/static_cell/2.1.0/static_cell/macro.make_static.html
@@ -58,27 +52,17 @@ async fn main(spawner: Spawner) -> ! {
     let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
     let peripherals = esp_hal::init(config);
 
-    // COEX needs more RAM - add some more
-    #[cfg(feature = "esp32")]
-    {
-        esp_alloc::heap_allocator!(#[ram(reclaimed)] size: 96 * 1024);
-        esp_alloc::heap_allocator!(size: 24 * 1024);
-    }
-    #[cfg(not(feature = "esp32"))]
-    {
-        esp_alloc::heap_allocator!(#[ram(reclaimed)] size: 64 * 1024);
-        esp_alloc::heap_allocator!(size: 64 * 1024);
-    }
+    esp_alloc::heap_allocator!(#[ram(reclaimed)] size: 64 * 1024);
+    esp_alloc::heap_allocator!(size: 36 * 1024);
 
     let timg0 = TimerGroup::new(peripherals.TIMG0);
     let sw_int = SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
     esp_rtos::start(timg0.timer0, sw_int.software_interrupt0);
 
-    let bluetooth = peripherals.BT;
-    let ble_controller = BleConnector::new(bluetooth, Default::default()).unwrap();
-
     let (controller, interfaces) =
         esp_radio::wifi::new(peripherals.WIFI, Default::default()).unwrap();
+
+    println!("CSI data will be printed shortly ...");
 
     let wifi_interface = interfaces.station;
 
@@ -98,9 +82,6 @@ async fn main(spawner: Spawner) -> ! {
     spawner.spawn(connection(controller)).ok();
     spawner.spawn(net_task(runner)).ok();
 
-    let mut rx_buffer = [0; 4096];
-    let mut tx_buffer = [0; 4096];
-
     loop {
         if stack.is_link_up() {
             break;
@@ -117,48 +98,8 @@ async fn main(spawner: Spawner) -> ! {
         Timer::after(Duration::from_millis(500)).await;
     }
 
-    core::mem::drop(ble_controller);
-    println!("Dropped BLE controller");
-
     loop {
         Timer::after(Duration::from_millis(1_000)).await;
-
-        let mut socket = TcpSocket::new(stack, &mut rx_buffer, &mut tx_buffer);
-
-        socket.set_timeout(Some(embassy_time::Duration::from_secs(10)));
-
-        let remote_endpoint = (Ipv4Addr::new(142, 250, 185, 115), 80);
-        println!("connecting...");
-        let r = socket.connect(remote_endpoint).await;
-        if let Err(e) = r {
-            println!("connect error: {:?}", e);
-            continue;
-        }
-        println!("connected!");
-        let mut buf = [0; 1024];
-        loop {
-            use embedded_io_async::Write;
-            let r = socket
-                .write_all(b"GET / HTTP/1.0\r\nHost: www.mobile-j.de\r\n\r\n")
-                .await;
-            if let Err(e) = r {
-                println!("write error: {:?}", e);
-                break;
-            }
-            let n = match socket.read(&mut buf).await {
-                Ok(0) => {
-                    println!("read EOF");
-                    break;
-                }
-                Ok(n) => n,
-                Err(e) => {
-                    println!("read error: {:?}", e);
-                    break;
-                }
-            };
-            println!("{}", core::str::from_utf8(&buf[..n]).unwrap());
-        }
-        Timer::after(Duration::from_millis(3000)).await;
     }
 }
 
@@ -184,6 +125,13 @@ async fn connection(mut controller: WifiController<'static>) {
             println!("Starting wifi");
             controller.start_async().await.unwrap();
             println!("Wifi started!");
+
+            let csi = CsiConfig::default();
+            controller
+                .set_csi(csi, |data: esp_radio::wifi::csi::WifiCsiInfo| {
+                    println!("rssi: {:?} rate: {}", data.rssi(), data.rate());
+                })
+                .unwrap();
 
             println!("Scan");
             let scan_config = ScanConfig::default().with_max(10);